feat : 네비게이션바 메뉴 카테고리 추가
