+++ src/components/Home.js
@@ -1,0 +1,8 @@
+function Home() {
+  return (
+    <div>
+      <h1>홈</h1>
+    </div>
+  );
+}
+export default Home;
